+++ app.js
@@ -1,67 +1,24 @@
 const express = require('express');
-const bodyParser = require('body-parser');
-const mysql = require('mysql');
-
-const db = mysql.createConnection({
-    host     : 'localhost',
-    user     : 'admin',
-    password : 'student',
-    database : 'ninjatasker'                                                                            
-
-});
-
-db.connect(function(err){
-    if (err) throw err;
-    console.log("DB is connected ...");
-});
-
-
-const urlEncoded = bodyParser.urlencoded({extended: false})
-
-const dummyData = [{taskItem: "Work on my portfolio" },{taskItem: "Code and watch anime"},{taskItem: "Sleep"}];
+const routes = require('./routes');
+const database = require('./models/todo')
 
 // setting up
 const app = express();
+
+// adding context to our request
+app.use( (req, res, next) => {
+    req.context = { db: database }
+    next();
+} );
+
+// routing manager
+app.use(routes);
 
 // setting template engine
 app.set("view engine","ejs");
 
 // use middle ware to serve static files
 app.use(express.static('./public'));
-
-// ############### ROUTES ##############
-
-// Get for tasks: returns all tasks
-app.get('/', (req, res) => {
-    let sql = 'SELECT * FROM task';
-    db.query(sql, function (err, results) {
-        if (err) throw err;
-        // rendering tasks view and passing taskToDo data
-        res.render('tasks', {taskToDo: results});
-    });
-});
-
-// Post for tasks: posting a task
-app.post('/tasks', urlEncoded, (req, res) => {
-    let task = req.body
-    let sql = 'INSERT INTO task SET ?';
-    db.query(sql, task, function (err, results) {
-        if (err) throw err;
-        // rendering tasks view and passing taskToDo data
-        console.log(results)
-        res.redirect('/')
-    });
-});
-
-// Delete for task: deleting specify task
-app.delete("/tasks/:id", (req, res) => {
-    let sql = 'DELETE  FROM task WHERE ID=' + req.params.id;
-    db.query(sql,(err, result) =>{
-        if(err) throw err;
-        console.log(result);
-        res.json(result)
-    })
-});
 
 // Server Test
 app.listen(3000, function(err){
